Refactor: formatDate 로직 분리
